(hel-445) Composant dans un bloc de test
